+++ src/main/java/beyond/samdasoo/sales/repository/SalesRepository.java
@@ -3,5 +3,5 @@ package beyond.samdasoo.sales.repository;
 import beyond.samdasoo.sales.entity.Sales;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SalesRepository extends JpaRepository<Sales, Integer> {
+public interface SalesRepository extends JpaRepository<Sales, Long> {
 }
